+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This project (loosely) adheres to [Semantic Versioning](https://semver.org/spec/
 
 ## Version 1.4.1 - Unreleased
 
+### Fixes
+* Cleanup in `util_import`
+
 
 ## Version 1.4.0 - Released 2025-08-16